basic http and fetch examples working
