+++ examples/with-fetch/package.json
@@ -1,16 +1,14 @@
 {
   "name": "ingest-with-fetch",
   "version": "1.0.0",
-  "description": "A simple boilerplate for using Ingest with Vercel.",
+  "description": "A simple boilerplate for using Ingest with fetch API.",
   "private": true,
   "scripts": {
-    "generate": "yarn build:tsc && yarn build:files",
-    "build:tsc": "tsc",
-    "build:files": "ts-node src/scripts/build.ts",
-    "dev": "ts-node src/scripts/server.ts"
+    "build": "tsc",
+    "dev": "ts-node src/server.ts"
   },
   "dependencies": {
-    "@stackpress/ingest-vercel": "0.2.14"
+    "@stackpress/ingest": "0.2.14"
   },
   "devDependencies": {
     "@types/node": "22.9.3",
