--- examples/with-fetch/src/routes/error.ts
@@ -1,14 +1,0 @@
-import { Context, Response, Exception } from '@stackpress/ingest';
-
-export default function ErrorResponse(req: Context, res: Response) {
-  try {
-    throw Exception.for('Not implemented');
-  } catch (e) {
-    const error = e as Exception;
-    res.setError({ 
-      code: error.code, 
-      error: error.message,
-      stack: error.trace()
-    });
-  }
-};
